move frets lib to lib/frets and remove some dead code
