+++ src/frets/system/pentatonic.js
