+++ src/lib/NoteIcon.svelte
@@ -1,0 +1,4 @@
+<script>
+</script>
+
+<circle class="fret-note-background" dx="0" r="12" />
