+++ vite.config.js
@@ -2,5 +2,10 @@ import { sveltekit } from '@sveltejs/kit/vite';
 import { defineConfig } from 'vite';
 
 export default defineConfig({
-  plugins: [sveltekit()]
+  plugins: [sveltekit()],
+  resolve: {
+    alias: {
+      '$frets': '/src/frets'
+    }
+  }
 });
